Rename tolerant cache to ephemeral cache
