+++ wcfsetup/install/files/lib/system/box/WhoWasOnlineBoxController.class.php
@@ -6,8 +6,8 @@ use wcf\data\DatabaseObject;
 use wcf\data\user\online\UserOnline;
 use wcf\data\user\online\UsersOnlineList;
 use wcf\data\user\UserProfile;
+use wcf\system\cache\ephemeral\WhoWasOnlineCache;
 use wcf\system\cache\runtime\UserProfileRuntimeCache;
-use wcf\system\cache\tolerant\WhoWasOnlineCache;
 use wcf\system\event\EventHandler;
 use wcf\system\WCF;
 use wcf\util\DateUtil;
