+++ wcfsetup/install/files/lib/system/user/UserBirthdayCache.class.php
@@ -29,7 +29,7 @@ class UserBirthdayCache extends SingletonFactory
     protected function loadMonth(int $month): void
     {
         if (!\array_key_exists($month, $this->birthdays)) {
-            $this->birthdays[$month] = (new \wcf\system\cache\tolerant\UserBirthdayCache($month))->get();
+            $this->birthdays[$month] = (new \wcf\system\cache\ephemeral\UserBirthdayCache($month))->get();
 
             $data = [
                 'birthdays' => $this->birthdays,
